% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_animations.R
\name{available_animations}
\alias{available_animations}
\title{Displaying the Available Animations}
\usage{
available_animations()
}
\value{
a list of the available animations by type of animation
}
\description{
The function displays a list of the available animations from the Animate.css library
}
\examples{
\donttest{


available_animations()

}

}
