% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gomo.R
\name{use_gomo}
\alias{use_gomo}
\title{Enable the animate.css library}
\usage{
use_gomo()
}
\value{
called for the side effect of activating the animate.css library
}
\description{
The function activates the capabilities of the animate.css library using its CDN. It should be run at the beginning of the document.
}
\examples{
\dontrun{

# Put the function at the beginning of the document

use_gomo()



}
}
