% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gomo_gomo_no.R
\name{gomo_gomo_no}
\alias{gomo_gomo_no}
\title{Generate Animated Text}
\usage{
gomo_gomo_no(
  text,
  level = "#",
  animation = "bounce",
  repeating = 1,
  delay = 0,
  duration = "slow",
  color = "black"
)
}
\arguments{
\item{text}{The text that should be animated}

\item{level}{The Markdown level of the text, available are : #, ##, ###, ####, #####, ######, text. Default to "#"}

\item{animation}{The type of animation. Default to "bounce".}

\item{repeating}{Number of times the animation is repeated (from 1 to 3). Default to 1}

\item{delay}{Time before triggering the animation (in seconds). Default to 0}

\item{duration}{The speed of the animation, available choices are: "slow", "slower", "fast", "faster". Default to "slow"}

\item{color}{the color of the animated text}
}
\value{
animated text
}
\description{
The function animates text using the Animate.css library
}
\examples{
\donttest{

# Put use_gomo() at the beginning of your Rmd Document or Shiny App (UI)
use_gomo()

# Then call gomo_gomo_no() wherever you want to animate your text

gomo_gomo_no(text = "I Love Chocolate !!!!", level = "###", animation = "tada")

}



}
